requirements.txt
awscli==1.16.145
boto3==1.9.134
Flask== 1.0.2
flask_sqlalchemy==2.3.2
numpy==1.16.3
pandas==0.24.2
pickleshare==0.7.5
pyaml==18.11.0
PyYAML==3.13
python-dateutil==2.7.5
pymysql==0.9.3
pytest==4.0.1
requests==2.21.0
scikit-learn==0.20.3
SQLAlchemy==1.3.1
